pyproject.toml
[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[project]
name = "tadv"
version = "0.11.0"
description = "Task-aware data validation constraint generation (TaDV)."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.6.0",
  "pydantic-settings>=2.2.0",
  "pyyaml>=6.0.0",
  "typing-extensions>=4.10.0; python_version < '3.12'",
]

[project.optional-dependencies]
deequ = [
  "numpy>=1.26.0",
  "pandas>=2.0.0",
  "pyspark>=3.5.0,<4.0.0",
  "pydeequ>=1.4.0",
]
duckdb = [
  "duckdb>=1.0.0",
]
gx = [
  "great_expectations>=1.3.0",
  "pandas>=2.0.0",
]
pandas = [
  "pandas>=2.0.0",
]
polars = [
  "polars>=0.20.0",
]
dspy = [
  "dspy-ai==3.1.2",
  "google-generativeai>=0.8.0",  # Required for Gemini models via litellm
]
api = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "python-multipart>=0.0.20",
  "slowapi>=0.1.9",
]

[dependency-groups]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.6.0",
  "mypy>=1.10.0",
  "types-PyYAML>=6.0.0",
  "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/tadv"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
